+++ src/components/fullscreen.vue
@@ -1,18 +1,21 @@
 
 <script>
   export default {
+    render: function () {
+      return '';
+    },
     computed: {
       callee: function () {
         return this.$store.state.callee;
       },
-      fullScreen: function () {
-        return this.$store.state.fullScreen;
+      fullscreen: function () {
+        return this.$store.state.fullscreen;
       }
     },
     watch: {
       'callee': function (val) {
-        if(val.name === 'fullScreen'){
-          this.$store.dispatch('setFullScreen', !this.fullScreen);
+        if(val.name === 'fullscreen'){
+          this.$store.dispatch('setFullScreen', !this.fullscreen);
         }
       }
     }