+++ src/plugins/emoji.vue
@@ -28,10 +28,7 @@
     <div class="ve-pop-body">
       <div class="wrap" @click="insertItem">
         <a href="javascript:;" draggable="false" v-for="item in arr">
-          <img
-          :src="parseSrc(item)"
-          :alt="twemoji.convert.fromCodePoint(item)" 
-          >
+          <img :src="parseSrc(item)" :alt="convert(item)">
         </a>
       </div>
     </div>
@@ -53,16 +50,19 @@
   export default {
     data () {
       return {
-        twemoji,
         arr: [
-          "1f600", "1f601", "1f602", "1f923", "1f603", "1f604", "1f605", "1f606", "1f609", "1f60a", "1f60b", "1f60e", "1f60d", "1f618",
-          "1f617", "1f619", "1f61a", "263a", "1f642", "1f917", "1f914", "1f610", "1f611", "1f636", "1f644", "1f60f", "1f623", "1f625",
-          "1f62e", "1f910", "1f62f", "1f62a", "1f62b", "1f634", "1f60c", "1f913", "1f61b", "1f61c", "1f61d", "1f924", "1f612", "1f613",
-          "1f614", "1f615", "1f643", "1f911", "1f632", "2639", "1f641", "1f616", "1f61e", "1f61f", "1f624", "1f622", "1f62d", "1f626",
-          "1f627", "1f628", "1f629", "1f62c", "1f630", "1f631", "1f633", "1f635", "1f621", "1f620", "1f607", "1f920", "1f921", "1f925",
-          "1f637", "1f912", "1f915", "1f922", "1f927"
+          '1f600', '1f601', '1f602', '1f923', '1f603', '1f604', '1f605', '1f606', '1f609', '1f60a', '1f60b', '1f60e', '1f60d', '1f618',
+          '1f617', '1f619', '1f61a', '263a', '1f642', '1f917', '1f914', '1f610', '1f611', '1f636', '1f644', '1f60f', '1f623', '1f625',
+          '1f62e', '1f910', '1f62f', '1f62a', '1f62b', '1f634', '1f60c', '1f913', '1f61b', '1f61c', '1f61d', '1f924', '1f612', '1f613',
+          '1f614', '1f615', '1f643', '1f911', '1f632', '2639', '1f641', '1f616', '1f61e', '1f61f', '1f624', '1f622', '1f62d', '1f626',
+          '1f627', '1f628', '1f629', '1f62c', '1f630', '1f631', '1f633', '1f635', '1f621', '1f620', '1f607', '1f920', '1f921', '1f925',
+          '1f637', '1f912', '1f915', '1f922', '1f927'
         ],
-        btn: '<a href="javascript:;" title="Insert emoji" unselectable="on"><i class="icon-smile-o"></i></a>',
+        // btn: '<a href="javascript:;" title="Insert emoji" unselectable="on"><i class="icon-smile-o"></i></a>',
+        btn: {
+          title: 'Insert emoji',
+          className: 'icon-smile-o'
+        }
       }
     },
     computed: {
@@ -72,6 +72,9 @@
       }
     },
     methods: {
+      convert (code) {
+        return twemoji.convert.fromCodePoint(code);
+      },
       parseSrc (str) {
         let div = document.createElement('div');
         div.innerHTML = twemoji.parse(twemoji.convert.fromCodePoint(str));